Tried to fix error with ModelLoader, couldnt do it tho

+++ Dusk Engine - Core/src/maths/Vector3.cpp
@@ -1,6 +1,7 @@
 #include "Vector3.h"
+#include <memory>
 
-Dusk::Maths::Vector3::Vector3(float x, float y, float z)
+Dusk::Maths::Vector3::Vector3(double x, double y, double z)
 {
 	m_C[0] = x;
 	m_C[1] = y;
